--- product_backend_intellij/src/test/java/com/shopping/cosmos/DemoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.shopping.cosmos;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
