+++ src/main/java/com/shopping/cosmos/DemoApplication.java
@@ -1,0 +1,13 @@
+package com.shopping.cosmos;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DemoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DemoApplication.class, args);
+	}
+
+}
