+++ src/main/java/com/shopping/cosmos/config/CORSFilter.java
@@ -1,0 +1,39 @@
+package com.shopping.cosmos.config;
+
+import javax.servlet.*;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public class CORSFilter implements Filter{
+
+//스프링 4.2버전 이후 불필요
+    @Override
+    public void doFilter(ServletRequest req, ServletResponse res, FilterChain chain) throws IOException, ServletException {
+
+        System.out.println("필터링이 진행중입니다..");
+        HttpServletResponse response = (HttpServletResponse)res;
+        response.setHeader("Access-Control-Allow-Origin","*");
+        response.setHeader("Access-Control-Allow-Credentials","true");
+        response.setHeader("Access-Control-Allow-Methods","POST,GET,PUT,OPTIONS,DELETE");
+        response.setHeader("Access-Control-Allow-Credentials","true");
+        response.setHeader("Access-Control-Allow-Max-Age","3600");
+        response.setHeader("Access-Control-Allow-Headers", "X-Requested-Width, Content-Type, " +
+                "Authorization, Origin, Accept, Access-Control-Request-Method, Access-Control-Request-Headers");
+
+        chain.doFilter(req,res);
+    }
+
+    @Override
+    public void init(FilterConfig filterConfig) throws ServletException {
+        Filter.super.init(filterConfig);
+    }
+
+    @Override
+    public void destroy() {
+        Filter.super.destroy();
+    }
+}
+
+
+
+
