+++ src/main/java/com/shopping/cosmos/vo/CartVO.java
@@ -1,0 +1,11 @@
+package com.shopping.cosmos.vo;
+
+import lombok.Data;
+
+@Data
+public class CartVO {
+    private int cart_id;
+    private String user_email;
+    private int product_seq;
+    private int amount;
+}
