+++ src/main/resources/application.properties
@@ -1,0 +1,9 @@
+#Oracle 설정
+spring.datasource.url=jdbc:oracle:thin:@localhost:1521:orcl
+spring.datasource.driver-class-name= oracle.jdbc.driver.OracleDriver
+spring.datasource.username=cosmos
+spring.datasource.password=cosmos
+
+#MyBatis 설정
+mybatis.mapper-locations=classpath*:mapper/*.xml
+
